Anathema startet wieder mit einem alten Report, aber die Additional Models scheinen nicht da zu sein

+++ Character_Reporting/src/net/sf/anathema/character/reporting/sheet/util/IPdfBoxEncoder.java
@@ -5,9 +5,9 @@ import net.sf.anathema.character.reporting.util.Bounds;
 import com.lowagie.text.pdf.PdfContentByte;
 
 public interface IPdfBoxEncoder {
-  public static final float HEADER_HEIGHT = 12;
-  public static final float ARCSPACE = HEADER_HEIGHT / 2;
-  public static final float ARC_SIZE = 2 * ARCSPACE;
+  public static final int HEADER_HEIGHT = 12;
+  public static final int ARCSPACE = HEADER_HEIGHT / 2;
+  public static final int ARC_SIZE = 2 * ARCSPACE;
 
   public void encodeContentBox(PdfContentByte directContent, Bounds bounds);
 }
